clauda

+++ database/migrations/2022_06_11_200847_create_trashes_table.php
@@ -5,29 +5,18 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
+
     public function up()
     {
         Schema::create('trashes', function (Blueprint $table) {
             $table->id();
             $table->string('table')->nullable();
-            $table->integer('user_id')->nullable();
-            $table->integer('table_id')->nullable();
-
-
+            $table->unsignedBigInteger('user_id')->nullable();
+            $table->unsignedBigInteger('table_id')->nullable();
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::dropIfExists('trashes');
